+++ src/main/java/com/infomatiq/jsi/rtree/Node.java
@@ -26,15 +26,15 @@ import com.infomatiq.jsi.Rectangle;
  */
 public class Node {
   int nodeId = 0;
-  float mbrMinX = Float.MAX_VALUE;
-  float mbrMinY = Float.MAX_VALUE;
-  float mbrMaxX = -Float.MAX_VALUE;
-  float mbrMaxY = -Float.MAX_VALUE;
+  double mbrMinX = Double.MAX_VALUE;
+  double mbrMinY = Double.MAX_VALUE;
+  double mbrMaxX = -Double.MAX_VALUE;
+  double mbrMaxY = -Double.MAX_VALUE;
 
-  float[] entriesMinX = null;
-  float[] entriesMinY = null;
-  float[] entriesMaxX = null;
-  float[] entriesMaxY = null;
+  double[] entriesMinX = null;
+  double[] entriesMinY = null;
+  double[] entriesMaxX = null;
+  double[] entriesMaxY = null;
 
   int[] ids = null;
   int level;
@@ -43,14 +43,14 @@ public class Node {
   Node(int nodeId, int level, int maxNodeEntries) {
     this.nodeId = nodeId;
     this.level = level;
-    entriesMinX = new float[maxNodeEntries];
-    entriesMinY = new float[maxNodeEntries];
-    entriesMaxX = new float[maxNodeEntries];
-    entriesMaxY = new float[maxNodeEntries];
+    entriesMinX = new double[maxNodeEntries];
+    entriesMinY = new double[maxNodeEntries];
+    entriesMaxX = new double[maxNodeEntries];
+    entriesMaxY = new double[maxNodeEntries];
     ids = new int[maxNodeEntries];
   }
 
-  void addEntry(float minX, float minY, float maxX, float maxY, int id) {
+  void addEntry(double minX, double minY, double maxX, double maxY, int id) {
     ids[entryCount] = id;
     entriesMinX[entryCount] = minX;
     entriesMinY[entryCount] = minY;
@@ -66,7 +66,7 @@ public class Node {
   }
 
   // Return the index of the found entry, or -1 if not found
-  int findEntry(float minX, float minY, float maxX, float maxY, int id) {
+  int findEntry(double minX, double minY, double maxX, double maxY, int id) {
     for (int i = 0; i < entryCount; i++) {
       if (id == ids[i] &&
           entriesMinX[i] == minX && entriesMinY[i] == minY &&
@@ -80,10 +80,10 @@ public class Node {
   // delete entry. This is done by setting it to null and copying the last entry into its space.
   void deleteEntry(int i) {
     int lastIndex = entryCount - 1;
-    float deletedMinX = entriesMinX[i];
-    float deletedMinY = entriesMinY[i];
-    float deletedMaxX = entriesMaxX[i];
-    float deletedMaxY = entriesMaxY[i];
+    double deletedMinX = entriesMinX[i];
+    double deletedMinY = entriesMinY[i];
+    double deletedMaxX = entriesMaxX[i];
+    double deletedMaxY = entriesMaxY[i];
 
     if (i != lastIndex) {
       entriesMinX[i] = entriesMinX[lastIndex];
@@ -101,8 +101,8 @@ public class Node {
 
   // deletedMin/MaxX/Y is a rectangle that has just been deleted or made smaller.
   // Thus, the MBR is only recalculated if the deleted rectangle influenced the old MBR
-  void recalculateMBRIfInfluencedBy(float deletedMinX, float deletedMinY,
-      float deletedMaxX, float deletedMaxY) {
+  void recalculateMBRIfInfluencedBy(double deletedMinX, double deletedMinY,
+      double deletedMaxX, double deletedMaxY) {
     if (mbrMinX == deletedMinX || mbrMinY == deletedMinY
         || mbrMaxX == deletedMaxX || mbrMaxY == deletedMaxY) {
       recalculateMBR();
@@ -162,19 +162,19 @@ public class Node {
     return level;
   }
 
-  public float getMbrMinX() {
+  public double getMbrMinX() {
     return mbrMinX;
   }
 
-  public float getMbrMinY() {
+  public double getMbrMinY() {
     return mbrMinY;
   }
 
-  public float getMbrMaxX() {
+  public double getMbrMaxX() {
     return mbrMaxX;
   }
 
-  public float getMbrMaxY() {
+  public double getMbrMaxY() {
     return mbrMaxY;
   }
 
@@ -182,19 +182,19 @@ public class Node {
     return new Rectangle(mbrMinX, mbrMinY, mbrMaxX, mbrMaxY);
   }
 
-  public float getEntryMbrMinX(int index) {
+  public double getEntryMbrMinX(int index) {
     return entriesMinX[index];
   }
 
-  public float getEntryMbrMinY(int index) {
+  public double getEntryMbrMinY(int index) {
     return entriesMinY[index];
   }
 
-  public float getEntryMbrMaxX(int index) {
+  public double getEntryMbrMaxX(int index) {
     return entriesMaxX[index];
   }
 
-  public float getEntryMbrMaxY(int index) {
+  public double getEntryMbrMaxY(int index) {
     return entriesMaxY[index];
   }
